Refactor shared code in tests and tools.
* Move SubprocessLauncher out of test/, into tool/. It's imported by both.
* Move most top-level Directory objects out of utils.dart, into singular test
  files where they are used, mostly dartdoc_test.dart.
* Move `generatorContextFromArgv` from utils.dart to dartdoc_test.dart.
* Clean up set-up and tear-down methods in dartdoc_test.dart.
* Move most top-level PackageGraph code from utils.dart to
  model_special_cases_test.dart and model_test.dart.

Also type-annotate public API in warnings.dart.
