+++ test/model_special_cases_test.dart
@@ -10,6 +10,7 @@ library dartdoc.model_special_cases_test;
 
 import 'dart:io';
 
+import 'package:async/async.dart';
 import 'package:dartdoc/dartdoc.dart';
 import 'package:dartdoc/src/model/model.dart';
 import 'package:dartdoc/src/special_elements.dart';
@@ -20,6 +21,43 @@ import 'src/utils.dart' as utils;
 
 final String _platformVersionString = Platform.version.split(' ').first;
 final Version _platformVersion = Version.parse(_platformVersionString);
+
+final _testPackageGraphExperimentsMemo = AsyncMemoizer<PackageGraph>();
+Future<PackageGraph> get _testPackageGraphExperiments =>
+    _testPackageGraphExperimentsMemo.runOnce(() => utils.bootBasicPackage(
+            'testing/test_package_experiments', [], additionalArguments: [
+          '--enable-experiment',
+          'non-nullable',
+          '--no-link-to-remote'
+        ]));
+
+final _testPackageGraphGinormousMemo = AsyncMemoizer<PackageGraph>();
+Future<PackageGraph> get _testPackageGraphGinormous =>
+    _testPackageGraphGinormousMemo.runOnce(() => utils.bootBasicPackage(
+            'testing/test_package', [
+          'css',
+          'code_in_commnets',
+          'excluded'
+        ], additionalArguments: [
+          '--auto-include-dependencies',
+          '--no-link-to-remote'
+        ]));
+
+final _testPackageGraphSmallMemo = AsyncMemoizer<PackageGraph>();
+Future<PackageGraph> get _testPackageGraphSmall =>
+    _testPackageGraphSmallMemo.runOnce(() => utils.bootBasicPackage(
+        'testing/test_package_small', [],
+        additionalArguments: ['--no-link-to-remote']));
+
+final _testPackageGraphSdkMemo = AsyncMemoizer<PackageGraph>();
+Future<PackageGraph> get _testPackageGraphSdk =>
+    _testPackageGraphSdkMemo.runOnce(_bootSdkPackage);
+
+Future<PackageGraph> _bootSdkPackage() async {
+  return PubPackageBuilder(
+          await utils.contextFromArgv(['--input', defaultSdkDir.path]))
+      .buildPackageGraph();
+}
 
 void main() {
   var sdkDir = defaultSdkDir;
@@ -46,17 +84,16 @@ void main() {
     Class b;
 
     setUpAll(() async {
-      lateFinalWithoutInitializer = (await utils.testPackageGraphExperiments)
+      lateFinalWithoutInitializer = (await _testPackageGraphExperiments)
           .libraries
           .firstWhere((lib) => lib.name == 'late_final_without_initializer');
-      nullSafetyClassMemberDeclarations = (await utils
-              .testPackageGraphExperiments)
+      nullSafetyClassMemberDeclarations = (await _testPackageGraphExperiments)
           .libraries
           .firstWhere((lib) => lib.name == 'nnbd_class_member_declarations');
-      optOutOfNullSafety = (await utils.testPackageGraphExperiments)
+      optOutOfNullSafety = (await _testPackageGraphExperiments)
           .libraries
           .firstWhere((lib) => lib.name == 'opt_out_of_nnbd');
-      nullableElements = (await utils.testPackageGraphExperiments)
+      nullableElements = (await _testPackageGraphExperiments)
           .libraries
           .firstWhere((lib) => lib.name == 'nullable_elements');
       b = nullSafetyClassMemberDeclarations.allClasses
@@ -283,7 +320,7 @@ void main() {
     PackageGraph ginormousPackageGraph;
 
     setUpAll(() async {
-      ginormousPackageGraph = await utils.testPackageGraphGinormous;
+      ginormousPackageGraph = await _testPackageGraphGinormous;
     });
 
     test('Verify that SDK libraries are not canonical when missing', () {
@@ -316,7 +353,7 @@ void main() {
     PackageGraph ginormousPackageGraph;
 
     setUpAll(() async {
-      ginormousPackageGraph = await utils.testPackageGraphGinormous;
+      ginormousPackageGraph = await _testPackageGraphGinormous;
     });
 
     test(
@@ -378,7 +415,7 @@ void main() {
     });
 
     test('Verify that packages without categories get handled', () async {
-      var packageGraphSmall = await utils.testPackageGraphSmall;
+      var packageGraphSmall = await _testPackageGraphSmall;
       expect(packageGraphSmall.localPackages.length, equals(1));
       expect(packageGraphSmall.localPackages.first.hasCategories, isFalse);
       var packageCategories = packageGraphSmall.localPackages.first.categories;
@@ -390,8 +427,8 @@ void main() {
     PackageGraph ginormousPackageGraph, sdkAsPackageGraph;
 
     setUpAll(() async {
-      ginormousPackageGraph = await utils.testPackageGraphGinormous;
-      sdkAsPackageGraph = await utils.testPackageGraphSdk;
+      ginormousPackageGraph = await _testPackageGraphGinormous;
+      sdkAsPackageGraph = await _testPackageGraphSdk;
     });
 
     group('test package', () {
@@ -424,7 +461,7 @@ void main() {
 
     group('test small package', () {
       test('does not have documentation', () async {
-        var packageGraphSmall = await utils.testPackageGraphSmall;
+        var packageGraphSmall = await _testPackageGraphSmall;
         expect(packageGraphSmall.defaultPackage.hasDocumentation, isFalse);
         expect(packageGraphSmall.defaultPackage.hasDocumentationFile, isFalse);
         expect(packageGraphSmall.defaultPackage.documentationFile, isNull);
@@ -471,7 +508,7 @@ void main() {
     Library dartAsyncLib;
 
     setUpAll(() async {
-      dartAsyncLib = (await utils.testPackageGraphSdk)
+      dartAsyncLib = (await _testPackageGraphSdk)
           .libraries
           .firstWhere((l) => l.name == 'dart:async');
       // Make sure the first library is dart:async
