+++ tool/grind.dart
@@ -12,7 +12,7 @@ import 'package:io/io.dart';
 import 'package:path/path.dart' as path;
 import 'package:yaml/yaml.dart' as yaml;
 
-import '../test/src/utils.dart';
+import 'subprocess_launcher.dart';
 
 void main([List<String> args]) => grind(args);
